ai/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
python-jose[cryptography]   
passlib[bcrypt] 
python-multipart

